// app/lib/types/FPLLeague.ts
// Generated by https://quicktype.io

export interface League {
    new_entries:       NewEntries;
    last_updated_data: string;
    league:            LeagueClass;
    standings:         NewEntries;
}

export interface LeagueClass {
    id:           number;
    name:         string;
    created:      string;
    closed:       boolean;
    max_entries:  null;
    league_type:  string;
    scoring:      string;
    admin_entry:  number;
    start_event:  number;
    code_privacy: string;
    has_cup:      boolean;
    cup_league:   null;
    rank:         null;
}

export interface NewEntries {
    has_next: boolean;
    page:     number;
    results:  Result[];
}

export interface Result {
    id:          number;
    event_total: number;
    player_name: string;
    rank:        number;
    last_rank:   number;
    rank_sort:   number;
    total:       number;
    entry:       number;
    entry_name:  string;
    player_first_name: string;
    player_last_name: string;
}